refactor: Sprint 3 — split vector_collection, reduce CC, migrate examples, translate docs
T3-1: Migrate 3 Rust examples from deprecated Collection API to typed
VectorCollection API (get_collection → get_vector_collection).

T3-2: Split vector_collection.rs (789 lines) into module directory:
  mod.rs (45), lifecycle.rs (86), accessors.rs (108),
  crud.rs (128), search.rs (413). All under 500 NLOC.

T3-4: Refactor 4 functions exceeding CC > 8:
  - bfs_traverse + bfs_traverse_reverse: shared BfsDirection enum +
    process_bfs_neighbors helper (eliminates 80% duplication)
  - traverse_dfs_config: extract expand_dfs_neighbors helper
  - remove_node_edges: extract purge_incoming/outgoing/label helpers
  3 functions already within CC limits (add_edge, wal_append_upsert,
  is_enabled) — skipped.

T3-5: Translate docs/guides/CONFIGURATION.md from French to English.
All code blocks preserved, technical terms unchanged.
